fix(dashmate): move filter to getServiceListFactory.js
